# backend/frontend_handoff/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Spatial System API",
    "description": "MVP API for natural geospatial query execution.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/planner/intent": {
      "post": {
        "summary": "Plan Intent",
        "description": "Plan natural geospatial query intent using LLM.\n\nDoes not execute plugins.",
        "operationId": "plan_intent_planner_intent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Plan Intent Planner Intent Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "summary": "Query Endpoint",
        "description": "Execute natural geospatial query.\n\nExpected body:\n    {\n        \"query\": \"...\",\n        \"inputs\": {...},\n        \"band_map\": {...},\n        \"request_id\": \"optional\",\n        \"user_context\": {...},\n        \"metadata\": {...},\n        \"min_score\": 0.01\n    }",
        "operationId": "query_endpoint_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/feedback": {
      "post": {
        "summary": "Feedback Endpoint",
        "description": "Submit user feedback for a previous request.\n\nExpected body:\n    {\n        \"request_id\": \"...\",\n        \"rating\": \"correct|incorrect|partial\",\n        \"issue_types\": [\"route_error\"],\n        \"expected_capability\": \"threshold_raster\",\n        \"expected_plugin_id\": \"raster_threshold\",\n        \"comment\": \"...\",\n        \"user_context\": {...}\n    }",
        "operationId": "feedback_endpoint_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Feedback Endpoint Feedback Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/planner/intent": {
      "post": {
        "summary": "Plan Intent",
        "description": "Plan natural geospatial query intent using LLM.\n\nDoes not execute plugins.",
        "operationId": "plan_intent_api_v1_planner_intent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Plan Intent Api V1 Planner Intent Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/query": {
      "post": {
        "summary": "Query Endpoint",
        "description": "Execute natural geospatial query.\n\nExpected body:\n    {\n        \"query\": \"...\",\n        \"inputs\": {...},\n        \"band_map\": {...},\n        \"request_id\": \"optional\",\n        \"user_context\": {...},\n        \"metadata\": {...},\n        \"min_score\": 0.01\n    }",
        "operationId": "query_endpoint_api_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feedback": {
      "post": {
        "summary": "Feedback Endpoint",
        "description": "Submit user feedback for a previous request.\n\nExpected body:\n    {\n        \"request_id\": \"...\",\n        \"rating\": \"correct|incorrect|partial\",\n        \"issue_types\": [\"route_error\"],\n        \"expected_capability\": \"threshold_raster\",\n        \"expected_plugin_id\": \"raster_threshold\",\n        \"comment\": \"...\",\n        \"user_context\": {...}\n    }",
        "operationId": "feedback_endpoint_api_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Feedback Endpoint Api V1 Feedback Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryRequest": {
        "properties": {
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "Natural-language geospatial query.",
            "examples": [
              "یک تحلیل ساده تستی انجام بده"
            ]
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs",
            "description": "Input datasets and references keyed by logical name."
          },
          "band_map": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Band Map",
            "description": "Optional raster band mapping, for example {'red': 3, 'nir': 4}."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Optional client-supplied request identifier."
          },
          "user_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "User Context",
            "description": "Optional user/session context."
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Optional request metadata."
          },
          "min_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Score",
            "description": "Optional minimum routing score."
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id",
            "description": "Optional project identifier."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "QueryRequest",
        "description": "Frontend-facing natural geospatial query request.\n\n`inputs` defaults to an empty object so the UI can start with a plain\nnatural-language query and attach datasets later when needed."
      },
      "QueryResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status, e.g. completed or failed."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "query_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Hash"
          },
          "ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ok"
          },
          "answer": {
            "title": "Answer"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "summary": {
            "title": "Summary"
          },
          "outputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Outputs"
          },
          "layers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Layers"
          },
          "map_layers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Map Layers"
          },
          "map": {
            "additionalProperties": true,
            "type": "object",
            "title": "Map"
          },
          "documents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Documents"
          },
          "files": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Files"
          },
          "reports": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Reports"
          },
          "artifacts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "warnings": {
            "items": {},
            "type": "array",
            "title": "Warnings"
          },
          "errors": {
            "items": {},
            "type": "array",
            "title": "Errors"
          },
          "next_actions": {
            "items": {},
            "type": "array",
            "title": "Next Actions"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "confidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Confidence"
          },
          "audit_ref": {
            "additionalProperties": true,
            "type": "object",
            "title": "Audit Ref"
          },
          "trace": {
            "items": {},
            "type": "array",
            "title": "Trace"
          },
          "steps": {
            "items": {},
            "type": "array",
            "title": "Steps"
          },
          "structured_error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Error"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "QueryResponse",
        "description": "Stable frontend-facing query response contract.\n\nThe backend may include additional diagnostic fields. The fields below are\nthe stable surface the UI can rely on."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}